Expand capability of mapps.directions and mapps.weather
Add 'from' and 'to' directives to mapps.directions. Both support city names
consisting of multiple words. The order of the directives is not important.
The current location will be used if no 'from' directive is present.

Add 'in' directive to mapps.weather. If present the given city will be used
instead of the current location. The city name may consist of multiple words.