Add priorities and completion progress to todo list
With 'priority <index> <level>' an arbitrary priority can be assigned
to each entry in the todo list. This priority ist the first key the list
will be sorted by. Priorities higher than 100 are colored red and ones
higher then 50 are yellow.

With the optional argument for 'complete' the progress of a task can be set.
This is the second key the todo list will be sorted by.