Lexical similarity (#69)
* Add string similarity functions.

Determine the best matching string to a given sentence, or find the best
fitting word in a list for your target word.

This function is used in reminders to remove entries by description and
not just the index.

* Use lexical similarity functions for trigger word determination in reminders.

Determine the course of action by using lexical similarity and a map between
trigger words and handler functions.

* Change Todo list command parsing to lexical similarity.

Move functions for lexical similarity into own utilities package.

* Move date and number parsing into utility library.

* Rearrange trigger list

* Expand lexical similarity matching for todo list.

Split handler for adding into different functions.
Trigger actions with multiple key words and sentences. When a suitable action
is found, the trigger sentence with most matched words is selected. Every matched
word is removed from the data string.

* Refactor handlers and trigger registration.

Add support for minimum argument count.

* Add lexical similarity to priority command.

* Divide reactions in reminder into single functions and use lexical similarity for trigger search.

* Remove unused imports and move print to GeneralUtilities

* Add documentation

* Add verbosity if not enough arguments

* Use lower case words for scoring

* Add lexical similarity summary NEW_FUNCTIONALITIES

* fix bug in help_todo