Update bulkresize.py file
- Add remove_backslash function. Some users of jarvis do not write the abolute path of the directory to the terminal but
they drag and drop the folder with images to the terminal. If the name of the folder has spaces amd the user do the above thing,
then the path to the directory looks like this /..path to directory/directory\ name/. This is not a valid path even if the directory exists.

- remove_backslash function simply removes the backslashes from the path and replace them with space in order to convert the path to
a valid path destination. So if the user gives a path like this /..path../directory\ name, which is an invalid path even if the directory exists,
remove backslash function will convert it to /..path../directory name which is a valid path name.