Fix directions_to.py and add an example
This fixes the directions command, when the user enters the start city before the destination city, in the command line argument.

It adds docstrings in the main() function of directions_to.py

It removes the wordIndex() function, which executes a command that has
already been executed and executes the index() method in the main() function.

It adds an example of directions with both start and destination city, in the converted.py.