Fix "near" command bug
This fixes the bug which occurs when the user asks for things near a specific location.

It adds Docstrings in the near_me.py

It enhances the code in near_me.py by removing the wordIndex() function and embedding its utility in the main(). Also changes the if statement to be more accurate.