Adding Zork Plugin
Please note the actual game content is not mine, and I just ported it over for Jarvis. I will eventually expand the content (possibly), and may even make a MUD out of it if anyone would be interested.

Also should Include the License into the code itself, or keep it a separate file?