Update translation.py to work with the newer googletrans module
Fixing a bug in the translation plugin where it is not working with the new default version of googletrans .

Also implementing some new features :

1. Adding the word auto as a source language for automatic language detection
2. Adding the ability to add a path as text input for .txt file translation*
3. Adding the ability to save the output to a file**

*The maximum number of characters the API can handle in one request is 15k, so I am implementing the ability to break the text 
into smaller chunks and translate them separately. The separation is done at a new line character for minimum semantic loss.

**The file is at a default path under the Jarvis directory and also for the case where the input is a file with more than 10k characters 
the output is stored in the file instead of being printed in the console as it would clutter it.