Fix playsound installation on Python 3.12+
On Python 3.12 and newer, the standard playsound package fails to build. This commit adds an environment marker in the requirements file so that for Python versions below 3.12, the normal playsound package is installed, and for Python 3.12 or above, the forked version (which includes fixes for newer Python versions) is installed. This change should resolve Issue #1235.